pdf-extractor-app/requirements.txt
streamlit==1.45.1
pandas==2.3.0
PyPDF2==3.0.1
openpyxl==3.1.5
openai==1.12.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-openai==0.0.2 